chore: moved OpenAPI generated Java DRC client code to separate package
