+++ src/main/java/net/atos/client/zgw/zrc/model/Zaak.java
@@ -20,7 +20,7 @@ import jakarta.json.bind.annotation.JsonbTransient;
 
 import org.apache.commons.collections4.CollectionUtils;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.shared.model.Archiefnominatie;
 import net.atos.client.zgw.zrc.model.generated.Opschorting;
 import net.atos.client.zgw.zrc.model.generated.ZaakKenmerk;
