+++ src/main/java/net/atos/client/zgw/zrc/model/ZaakListParameters.java
@@ -15,7 +15,7 @@ import jakarta.ws.rs.QueryParam;
 
 import org.apache.commons.collections4.CollectionUtils;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.shared.model.AbstractListParameters;
 import net.atos.client.zgw.shared.model.Archiefnominatie;
 import net.atos.client.zgw.ztc.model.AardVanRol;
