+++ src/main/java/net/atos/zac/app/informatieobjecten/EnkelvoudigInformatieObjectDownloadService.java
@@ -23,7 +23,7 @@ import jakarta.inject.Inject;
 import jakarta.ws.rs.core.StreamingOutput;
 
 import net.atos.client.zgw.drc.DRCClientService;
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.zrc.ZRCClientService;
 import net.atos.client.zgw.zrc.model.ZaakInformatieobject;
