+++ src/main/java/net/atos/zac/app/informatieobjecten/model/RESTEnkelvoudigInformatieObjectVersieGegevens.java
@@ -8,7 +8,7 @@ package net.atos.zac.app.informatieobjecten.model;
 import java.time.LocalDate;
 import java.util.UUID;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.zac.app.configuratie.model.RESTTaal;
