+++ src/main/java/net/atos/zac/app/productaanvragen/InboxProductaanvragenRESTService.java
@@ -27,7 +27,7 @@ import jakarta.ws.rs.core.Response;
 import org.apache.commons.collections4.CollectionUtils;
 
 import net.atos.client.zgw.drc.DRCClientService;
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.zac.aanvraag.InboxProductaanvraagService;
 import net.atos.zac.aanvraag.model.InboxProductaanvraagListParameters;
 import net.atos.zac.aanvraag.model.InboxProductaanvraagResultaat;
