+++ src/main/java/net/atos/zac/documentcreatie/model/DocumentCreatieGegevens.java
@@ -5,7 +5,7 @@
 
 package net.atos.zac.documentcreatie.model;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.zrc.model.Zaak;
 import net.atos.client.zgw.ztc.model.generated.InformatieObjectType;
