+++ src/main/java/net/atos/zac/documenten/OntkoppeldeDocumentenService.java
@@ -28,7 +28,7 @@ import jakarta.transaction.Transactional;
 
 import org.apache.commons.lang3.StringUtils;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.shared.util.DateTimeUtil;
 import net.atos.client.zgw.zrc.model.Zaak;
 import net.atos.zac.authentication.LoggedInUser;
