+++ src/main/java/net/atos/zac/mail/model/Bronnen.java
@@ -2,7 +2,7 @@ package net.atos.zac.mail.model;
 
 import org.flowable.task.api.TaskInfo;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.zrc.model.Zaak;
 
 public class Bronnen {
