+++ src/main/java/net/atos/zac/mailtemplates/MailTemplateHelper.java
@@ -50,7 +50,7 @@ import net.atos.client.brp.model.VerblijfadresBuitenland;
 import net.atos.client.brp.model.VerblijfplaatsBuitenland;
 import net.atos.client.kvk.KVKClientService;
 import net.atos.client.kvk.zoeken.model.ResultaatItem;
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.shared.ZGWApiService;
 import net.atos.client.zgw.zrc.ZRCClientService;
 import net.atos.client.zgw.zrc.model.BetrokkeneType;
