+++ src/main/java/net/atos/zac/signalering/event/SignaleringEventUtil.java
@@ -12,7 +12,7 @@ import java.util.Set;
 
 import org.flowable.task.api.TaskInfo;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.zrc.model.Zaak;
 import net.atos.zac.identity.model.User;
 import net.atos.zac.notificaties.Action;
