+++ src/main/java/net/atos/zac/zoeken/model/zoekobject/DocumentZoekObject.java
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 
 import org.apache.solr.client.solrj.beans.Field;
 
-import net.atos.client.zgw.drc.model.EnkelvoudigInformatieObject;
+import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.zac.zoeken.model.DocumentIndicatie;
 import net.atos.zac.zoeken.model.ZoekObject;
 import net.atos.zac.zoeken.model.index.ZoekObjectType;
